// src/notion/properties/types.ts
export type PropertyType =
  | 'checkbox'
  | 'created_by'
  | 'created_time'
  | 'date'
  | 'email'
  | 'file'
  | 'formula'
  | 'last_edited_by'
  | 'last_edited_time'
  | 'multi_select'
  | 'number'
  | 'person'
  | 'phone_number'
  | 'relation'
  | 'rollup'
  | 'select'
  | 'text'
  | 'title'
  | 'url'

export const PROPERTY_TYPES = {
  checkbox: 'checkbox',
  createdBy: 'created_by',
  createdTime: 'created_time',
  date: 'date',
  email: 'email',
  file: 'file',
  formula: 'formula',
  lastEditedBy: 'last_edited_by',
  lastEditedTime: 'last_edited_time',
  multiSelect: 'multi_select',
  number: 'number',
  person: 'person',
  phoneNumber: 'phone_number',
  relation: 'relation',
  rollup: 'rollup',
  select: 'select',
  text: 'text',
  title: 'title',
  url: 'url'
}